refactor
removed comented code and made changes in addcart component
