+++ src/components/Index.js
@@ -1,0 +1,6 @@
+import { Navbar } from '../../components/Navbar/Navbar';
+import {Footer} from '../../components/Footer/Footer';
+import {AddToCart} from '../../components/AddToCart/AddToCart';
+import { CategoriesNavbar } from '../../components/CategoriesNavbar/CategoriesNavbar';
+import { LatestCart } from '../../components/LatestCart/LatestCart';
+import { Filter} from '../../components/Filter/Filter';
