+++ src/components/Navbar/Navbar.js
@@ -1,5 +1,9 @@
 import React from 'react'
+import { Link } from 'react-router-dom'
+import { Homepage } from '../../pages/Homepage/Homepage'
 import './Navbar.css'
+
+
 
 
 export const Navbar = () => {
@@ -7,7 +11,7 @@ export const Navbar = () => {
     <div>
         <nav className="navigation">
             <ul className="left_Brand">
-                <a href="./homepage.html">Ethinic Wear</a>
+                <Link to="/">Ethinic Wear</Link>
             </ul>
             <ul className="right_Brand">
                 <li><a className="active" href="./login.html">Login
