+++ src/pages/Cart/Cart.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Navbar } from '../components/Navbar/Navbar'
+import { Navbar } from '../../components/Navbar/Navbar'
 
 
 export const Cart = () => {
