--- src/pages/Productpage.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Navbar } from '../components/Navbar/Navbar'
-
-
-export const Productpage = () => {
-  return (
-    <div>
-     <Navbar/>  
-     
-    </div>
-  )
-}
