--- src/pages/Wishlist.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import { CategoriesNavbar } from '../components/CategoriesNavbar/CategoriesNavbar'
-import { Navbar } from '../components/Navbar/Navbar'
-
-
-export const Wishlist = () => {
-  return (
-    <div>
-     <Navbar/>  
-     <CategoriesNavbar/>
-    </div>
-  )
-}
