+++ src/pages/Wishlist/Wishlist.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { CategoriesNavbar } from '../../components/CategoriesNavbar/CategoriesNavbar'
+import { Navbar } from '../../components/Navbar/Navbar'
+
+
+export const Wishlist = () => {
+  return (
+    <div>
+     <Navbar/>  
+     <CategoriesNavbar/>
+    </div>
+  )
+}
